Estimate length-at-age parameters and growth curves for Elasmobranchs

A multi-model growth estimation approach is applied to length-at-age data. Three models can be applied which include the von Bertalanffy (VB), logistic (Log) and Gompertz (Gom) models. AIC values and weights are calculated. The outputs will return a list of model parameter estimates and will either print a plot to the screen or output the length-at-age estimates as part of the list.Use of this function should cite Smart et al. (2016).

Estimate_Growth(
  data,
  models = c("VB", "Log", "Gom"),
  Birth.Len = NULL,
  correlation.matrix = FALSE,
  n.bootstraps = 1000,
  plots = T,
  Max.Age = NULL,
  plot.legend = T
)

Arguments

data

a data frame which includes 'Age' and 'Length - ideally with these names but the function will except some variation to these

models

a vector of models to be fitted. These can include" VB", "Log" and "Gom". A subset can also be used

Birth.Len

The length-at-birth to be used for two parameter models. If a value is provided, two parameter models are automatically run

correlation.matrix

Should the correlation matrix of parameters be returned? This is the only object returned if TRUE.

n.bootstraps

The number of bootstraps performed for model 95 confidence intervals

plots

Should plots be printed to the screen. If FALSE then the model estimates and CI's are returned as an additional output

Max.Age

Specify the max age for bootstrapped confidence intervals to be produced over. Default is the max age in the data.

plot.legend

Do you want a legend for the different models on the plot

Value

Returns a list of parameter estimates with errors and AIC results. If plots is TRUE then a plot is printed to the screen. If plots is FALSE then the length-at-age estimates are returned as a list element

References

Smart et al. (2016) Multi-model approaches in shark and ray growth studies: strengths, weaknesses and the future. Fish and Fisheries. 17: 955-971https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12154