Estimate age-at-maturity

Age-at-maturity is estimated from binary maturity data using a logistic ogive. Two options are available depending on error structure. If binary data are used then a binomial error structure is required. If the user wishes to bin the data by age class then a quasi binomial error structure is needed with the data weighted by the sample size of each bin. This is handled automatically by the function.

Estimate_Age_Maturity(
  data,
  error.structure = "binomial",
  n.bootstraps = 1000,
  display.points = FALSE,
  return = "parameters"
)

Arguments

data

A dataframe that includes age and a binary maturity status (immature = 0 and mature = 1). Columns should be named "Age" and "Maturity" but the function is robust enough to accept some reasonable variations to these

error.structure

The distribution for the glm used to produce the logistic ogive. Must be either "binomial" for binary data or "quasi binomial" for binned maturity at age. Proportion mature at each age is automatically calculated within the function

n.bootstraps

Number of bootstrap iterations required to produce 95% confidence intervals about the logistic ogive

display.points

Should the raw data be plotted for the binomial model?

return

Either:

parameters

The estimated logistic parameters and their standard error (A50 and A95)

estimates

The logistic ogive predictions with 95 percent confidence intervals (useful for creating ones own plots)

plot

a ggplot object of the logistic ogive.

Value

Either:

parameters

a dataframe of the estimated logistic parameters and their standard error (A50 and A95)

estimates

a dataframe of logistic ogive predictions with 95 percent confidence intervals

plot

a ggplot object of the logistic ogive